validation  rentals

+++ src/controllers/rentals/postRentalsControllers.js
@@ -9,10 +9,27 @@ export async function addRentals(req,res){
     
     const {customerId ,gameId, daysRented} = req.body
 
+    if(daysRented<=0){
+        return res.sendStatus(400)
+    }
+
     let { rows: pricePerDay} = await connection.query(`
         SELECT "pricePerDay" FROM games  
         WHERE id = ($1);
     `,[gameId])
+
+    if(pricePerDay.length ===0){
+        return res.sendStatus(400)
+    }
+
+    let { rows: iscustomers} = await connection.query(`
+    SELECT * FROM customers  
+    WHERE id = ($1);
+`,[customerId])
+
+if(iscustomers.length ===0){
+    return res.sendStatus(400)
+}
 
     pricePerDay = pricePerDay[0].pricePerDay
     const originalPrice = pricePerDay*daysRented
@@ -33,15 +50,28 @@ export async function finishRentals(req, res){
 
     const id =  req.params.id
 
+    console.log(id)
+
 
 
         const { rows: dataRental} = await connection.query(`
-        SELECT "rentDate", "pricePerDay" 
+        SELECT "rentDate", "pricePerDay", "returnDate" 
         FROM rentals  
         JOIN games
         ON rentals."gameId" = games.id
         WHERE rentals.id = ($1);
     `,[id])
+
+    console.log(dataRental[0].returnDate)
+
+    if(dataRental.length===0){
+        return res.sendStatus(404)
+    }
+    if(dataRental[0].returnDate!==null){
+        return res.sendStatus(400)
+    }
+
+
 
 
     const rentDate = dataRental[0].rentDate
@@ -50,19 +80,15 @@ export async function finishRentals(req, res){
 
     const dayNow = dayjs().format('YYYY-MM-DD')
     const formatedDate = dayjs(dayNow);
-
-
     const delayDay =  formatedDate.diff(rentDate, 'day');
-
-    console.log(delayDay)
-    console.log(pricePerDay)
     const delayFee = delayDay*pricePerDay
     
     
     await connection.query(`
-    UPDATE rentals SET "returnDate"  = ($1), "delayFee" = ($2)
+    UPDATE rentals SET "returnDate" = ($1), "delayFee" = ($2)
+    WHERE rentals.id = ($3)
  
-        `, [dayNow, delayFee] )
+        `, [dayNow, delayFee, id] )
 
     res.sendStatus(200)
 }
@@ -70,6 +96,22 @@ export async function finishRentals(req, res){
 export async function deleteRentals(req, res){
 
  const id =  req.params.id
+
+
+ const { rows: dataRental} = await connection.query(`
+ SELECT "returnDate" 
+ FROM rentals  
+ WHERE rentals.id = ($1);
+`,[id])
+
+console.log(dataRental[0].returnDate)
+
+if(dataRental.length===0){
+ return res.sendStatus(404)
+}
+if(dataRental[0].returnDate===null){
+ return res.sendStatus(400)
+}
 
     connection.query(`
         DELETE FROM rentals
